+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Security`
+## [25.3.5] - 2021-02-09
+
+- SC-8610 - fix update route
+
+## [25.3.4] - 2021-02-08
+
+### Fixed
+
+- SC-8604 - fix delete event route
 
 ## [25.3.2] - 2021-02-02